requirements.txt
certifi==2018.4.16
chardet==3.0.4
defusedxml==0.5.0
Django==2.0.4
django-allauth==0.35.0
django-autoslug-iplweb==1.9.4.dev0
django-background-tasks==1.1.13
django-compat==1.0.15
django-model-utils==3.1.1
idna==2.6
jsonfield==2.0.2
oauthlib==2.0.7
Pillow==5.1.0
python3-openid==3.1.0
pytz==2018.4
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
urllib3==1.22
